Created Trip Controller
